requirements.txt
requests
beautifulsoup4
Flask
python-dotenv
